chore: refactor tests and address open todos
