+++ cypress/e2e/micro-session-workflow.cy.ts
@@ -1,0 +1,212 @@
+describe('Different micro-session workflows', () => {
+    beforeEach(() => {
+        cy.visit(Cypress.env('URL_LECTURER'));
+        cy.get('[data-cy="login-logo"]').should('exist');
+        cy.get('[data-cy="email-field"]').type(Cypress.env('LECTURER_EMAIL'));
+        cy.get('[data-cy="password-field"]').type(Cypress.env('LECTURER_PASSWORD'));
+        cy.get('[data-cy="submit-login"]').click();
+    }),
+
+
+    it('creates and publishes a micro session that should be visible to students', () => {
+        const randomNumber = Math.round(Math.random() * 1000);
+        const questionTitle = 'A Single Choice with solution' + randomNumber;
+        const question = 'Was ist die Wahrscheinlichkeit? ' + randomNumber;
+        const microSessionName = 'Test Micro-Session ' + randomNumber;
+        const microSessionDisplayName = 'Displayed Name ' + randomNumber;
+        const description = 'This is the official descriptioin of ' + randomNumber
+    
+        // set up question
+        cy.get('[data-cy="create-question"]').click();
+        cy.get('[data-cy="insert-question-title"]').click().type(questionTitle);
+        cy.get('[data-cy="insert-question-text"]').click().type(question);
+        cy.get('[data-cy="configure-sample-solution"]').click();
+        cy.get('[data-cy="insert-answer-field"]').click().type('50%');
+        cy.get('[data-cy="set-correctness"]').click({force: true});
+        cy.get('[data-cy="add-new-answer"]').click({force: true});
+        cy.get('[data-cy="insert-answer-field"]').eq(1).click().type('100%');
+        cy.get('[data-cy="save-new-question"]').click({force: true});
+        // 
+        // create a micro-session
+        cy.get('[data-cy="create-micro-session"]').click();
+        // 
+        // step 1
+        cy.get('[data-cy="insert-micro-session-name"]').click().type(microSessionName);
+        cy.get('[data-cy="insert-micro-session-display-name"]').click().type(microSessionDisplayName);
+        cy.get('[data-cy="insert-micro-session-description"]').click().type(description)
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        // step 2
+        cy.get('[data-cy="select-course"]').should('exist').contains("Testkurs"); 
+        cy.get('[data-cy="select-start-date"]').click().type("2023-01-01T18:00");
+        cy.get('[data-cy="select-end-date"]').click().type("2023-12-31T18:00");
+        cy.get('[data-cy="select-multiplier"]').should('exist').contains('Einfach (1x)'); 
+        cy.get('[data-cy="select-multiplier"]').click().siblings().eq(0).findByText('Doppelt (2x)').parent().click();
+        cy.get('[data-cy="select-multiplier"]').contains('Doppelt (2x)');
+        cy.get('[data-cy="next-or-submit"]').click()
+    
+        // step 3
+        const dataTransfer = new DataTransfer();
+        cy.get('[data-cy="question-block"]').contains(questionTitle).trigger('dragstart', {
+          dataTransfer
+        });
+        cy.get('[data-cy="drop-questions-here"]').trigger('drop', {
+          dataTransfer
+        });
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        cy.get('[data-cy="load-session-list"]').click();
+        cy.get('[data-cy="micro-session"]').contains(microSessionName);
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Draft');
+    
+        // publish a micro-session
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').siblings().children().get('[data-cy="publish-micro-session"]').contains('Micro-Session veröffentlichen').click();
+        cy.get('[data-cy="verify-publish-action"]').click()
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Published');
+
+        // sign in as student
+        cy.clearAllCookies();
+        cy.visit(Cypress.env('URL_STUDENT'));
+        cy.get('[data-cy="username-field"]').click().type(Cypress.env('STUDENT_USERNAME'));
+        cy.get('[data-cy="password-field"]').click().type(Cypress.env('STUDENT_PASSWORD'));
+        cy.get('[data-cy="submit-login"]').click();
+
+        cy.contains('[data-cy="micro-learnings"]', microSessionDisplayName).should('exist');
+        cy.findByText(microSessionDisplayName).click();
+        cy.get('[data-cy="start-micro-session"]').click();
+        cy.get('[data-cy="choice-option"]').eq(0).click();
+        // cy.get('[data-cy="send-answer"]').click(); //TODO: cypress command works but there is an unhandled promise rejection that causes the test to fail
+    }),
+
+    it('creates and publishes a future micro session that should not be visible to students', () => {
+        const randomNumber = Math.round(Math.random() * 1000);
+        const questionTitle = 'A Single Choice with solution' + randomNumber;
+        const question = 'Was ist die Wahrscheinlichkeit? ' + randomNumber;
+        const microSessionName = 'Test Micro-Session ' + randomNumber;
+        const microSessionDisplayName = 'Displayed Name ' + randomNumber;
+        const description = 'This is the official descriptioin of ' + randomNumber
+    
+        // set up question
+        cy.get('[data-cy="create-question"]').click();
+        cy.get('[data-cy="insert-question-title"]').click().type(questionTitle);
+        cy.get('[data-cy="insert-question-text"]').click().type(question);
+        cy.get('[data-cy="configure-sample-solution"]').click();
+        cy.get('[data-cy="insert-answer-field"]').click().type('50%');
+        cy.get('[data-cy="set-correctness"]').click({force: true});
+        cy.get('[data-cy="add-new-answer"]').click({force: true});
+        cy.get('[data-cy="insert-answer-field"]').eq(1).click().type('100%');
+        cy.get('[data-cy="save-new-question"]').click({force: true});
+         
+        // create a micro-session
+        cy.get('[data-cy="create-micro-session"]').click();
+         
+        // step 1
+        cy.get('[data-cy="insert-micro-session-name"]').click().type(microSessionName);
+        cy.get('[data-cy="insert-micro-session-display-name"]').click().type(microSessionDisplayName);
+        cy.get('[data-cy="insert-micro-session-description"]').click().type(description)
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        // step 2
+        cy.get('[data-cy="select-course"]').should('exist').contains("Testkurs"); 
+        cy.get('[data-cy="select-start-date"]').click().type("2024-01-01T18:00");
+        cy.get('[data-cy="select-end-date"]').click().type("2024-12-31T18:00");
+        cy.get('[data-cy="select-multiplier"]').should('exist').contains('Einfach (1x)'); 
+        cy.get('[data-cy="select-multiplier"]').click().siblings().eq(0).findByText('Doppelt (2x)').parent().click();
+        cy.get('[data-cy="select-multiplier"]').contains('Doppelt (2x)');
+        cy.get('[data-cy="next-or-submit"]').click()
+    
+        // step 3
+        const dataTransfer = new DataTransfer();
+        cy.get('[data-cy="question-block"]').contains(questionTitle).trigger('dragstart', {
+          dataTransfer
+        });
+        cy.get('[data-cy="drop-questions-here"]').trigger('drop', {
+          dataTransfer
+        });
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        cy.get('[data-cy="load-session-list"]').click();
+        cy.get('[data-cy="micro-session"]').contains(microSessionName);
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Draft');
+    
+        // publish a micro-session
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').siblings().children().get('[data-cy="publish-micro-session"]').contains('Micro-Session veröffentlichen').click();
+        cy.get('[data-cy="verify-publish-action"]').click()
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Published');
+
+        // sign in as student
+        cy.clearAllCookies();
+        cy.visit(Cypress.env('URL_STUDENT'));
+        cy.get('[data-cy="username-field"]').click().type(Cypress.env('STUDENT_USERNAME'));
+        cy.get('[data-cy="password-field"]').click().type(Cypress.env('STUDENT_PASSWORD'));
+        cy.get('[data-cy="submit-login"]').click();
+
+        cy.contains('[data-cy="micro-learnings"]', microSessionDisplayName).should('not.exist');
+    }),
+
+    it('creates and publishes a past micro session that should not be visible to students', () => {
+        const randomNumber = Math.round(Math.random() * 1000);
+        const questionTitle = 'A Single Choice with solution' + randomNumber;
+        const question = 'Was ist die Wahrscheinlichkeit? ' + randomNumber;
+        const microSessionName = 'Test Micro-Session ' + randomNumber;
+        const microSessionDisplayName = 'Displayed Name ' + randomNumber;
+        const description = 'This is the official descriptioin of ' + randomNumber
+    
+        // set up question
+        cy.get('[data-cy="create-question"]').click();
+        cy.get('[data-cy="insert-question-title"]').click().type(questionTitle);
+        cy.get('[data-cy="insert-question-text"]').click().type(question);
+        cy.get('[data-cy="configure-sample-solution"]').click();
+        cy.get('[data-cy="insert-answer-field"]').click().type('50%');
+        cy.get('[data-cy="set-correctness"]').click({force: true});
+        cy.get('[data-cy="add-new-answer"]').click({force: true});
+        cy.get('[data-cy="insert-answer-field"]').eq(1).click().type('100%');
+        cy.get('[data-cy="save-new-question"]').click({force: true});
+         
+        // create a micro-session
+        cy.get('[data-cy="create-micro-session"]').click();
+         
+        // step 1
+        cy.get('[data-cy="insert-micro-session-name"]').click().type(microSessionName);
+        cy.get('[data-cy="insert-micro-session-display-name"]').click().type(microSessionDisplayName);
+        cy.get('[data-cy="insert-micro-session-description"]').click().type(description)
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        // step 2
+        cy.get('[data-cy="select-course"]').should('exist').contains("Testkurs"); 
+        cy.get('[data-cy="select-start-date"]').click().type("2021-01-01T18:00");
+        cy.get('[data-cy="select-end-date"]').click().type("2021-12-31T18:00");
+        cy.get('[data-cy="select-multiplier"]').should('exist').contains('Einfach (1x)'); 
+        cy.get('[data-cy="select-multiplier"]').click().siblings().eq(0).findByText('Doppelt (2x)').parent().click();
+        cy.get('[data-cy="select-multiplier"]').contains('Doppelt (2x)');
+        cy.get('[data-cy="next-or-submit"]').click()
+    
+        // step 3
+        const dataTransfer = new DataTransfer();
+        cy.get('[data-cy="question-block"]').contains(questionTitle).trigger('dragstart', {
+          dataTransfer
+        });
+        cy.get('[data-cy="drop-questions-here"]').trigger('drop', {
+          dataTransfer
+        });
+        cy.get('[data-cy="next-or-submit"]').click();
+    
+        cy.get('[data-cy="load-session-list"]').click();
+        cy.get('[data-cy="micro-session"]').contains(microSessionName);
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Draft');
+    
+        // publish a micro-session
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').siblings().children().get('[data-cy="publish-micro-session"]').contains('Micro-Session veröffentlichen').click();
+        cy.get('[data-cy="verify-publish-action"]').click()
+        cy.findByText(microSessionName).parentsUntil('[data-cy="micro-session"]').contains('Published');
+
+        // sign in as student
+        cy.clearAllCookies();
+        cy.visit(Cypress.env('URL_STUDENT'));
+        cy.get('[data-cy="username-field"]').click().type(Cypress.env('STUDENT_USERNAME'));
+        cy.get('[data-cy="password-field"]').click().type(Cypress.env('STUDENT_PASSWORD'));
+        cy.get('[data-cy="submit-login"]').click();
+
+        cy.contains('[data-cy="micro-learnings"]', microSessionDisplayName).should('not.exist');
+    })
+})
